+++ internal/silk/silk.go
@@ -66,7 +66,7 @@ func clamp(low, in, high int32) int32 {
 	return in
 }
 
-func clampFloat(low, in, high float64) float64 {
+func clampFloat(low, in, high float32) float32 {
 	if in > high {
 		return high
 	} else if in < low {
@@ -77,9 +77,11 @@ func clampFloat(low, in, high float64) float64 {
 }
 
 // The sign of x, i.e.,
-//            ( -1,  x < 0
-//  sign(x) = <  0,  x == 0
-//            (  1,  x > 0
+//
+//	          ( -1,  x < 0
+//	sign(x) = <  0,  x == 0
+//	          (  1,  x > 0
+//
 // https://datatracker.ietf.org/doc/html/rfc6716#section-1.1.4
 func sign(x int) int {
 	switch {
@@ -92,12 +94,12 @@ func sign(x int) int {
 	}
 }
 
-//  The minimum number of bits required to store a positive integer n in
-//  binary, or 0 for a non-positive integer n.
+// The minimum number of bits required to store a positive integer n in
+// binary, or 0 for a non-positive integer n.
 //
-//                             ( 0,                 n <= 0
-//                   ilog(n) = <
-//                             ( floor(log2(n))+1,  n > 0
+//	          ( 0,                 n <= 0
+//	ilog(n) = <
+//	          ( floor(log2(n))+1,  n > 0
 func ilog(n int) int {
 	if n <= 0 {
 		return 0
